requirements.txt
pandas==2.2.3
Pillow==11.0.0
pygame==2.5.2
Requests==2.32.3